delete hibernate module from develeloper branch

+++ travel-company/travel-company-common/src/main/resources/db/migration/V3__Add_airport.sql
@@ -1,6 +1,6 @@
 CREATE TABLE airport
 (
-    id           int         NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    id           bigint      NOT NULL PRIMARY KEY AUTO_INCREMENT,
     airport_code VARCHAR(10) NOT NULL UNIQUE,
     country      VARCHAR(30) NOT NULL,
     city         VARCHAR(30) NOT NULL
