+++ travel-company/travel-company-common/src/main/resources/db/migration/V6__Add_routemap.sql
@@ -1,10 +1,10 @@
 CREATE TABLE routemap
 (
-    id                     int primary key AUTO_INCREMENT,
-    airline_id             int NOT NULL,
-    origin_airport_id      int NOT NULL,
-    destination_airport_id int NOT NULL,
-    direction_id           int NOT NULL,
+    id                     bigint primary key AUTO_INCREMENT,
+    airline_id             bigint NOT NULL,
+    origin_airport_id      bigint NOT NULL,
+    destination_airport_id bigint NOT NULL,
+    direction_id           bigint NOT NULL,
     CONSTRAINT airline_fk FOREIGN KEY (airline_id)
         REFERENCES airline (id) ON DELETE CASCADE,
     CONSTRAINT origin_airport_fk FOREIGN KEY (origin_airport_id)
